requirements.txt
Django==2.0.8
django-cors-headers==2.5.2
djangorestframework==3.9.2
Markdown==3.1
psycopg2==2.8.1
pytz==2019.1
